Move edition capability logic to utils, leave the row type in types
types/edition.ts had grown into the largest runtime module under types/ --
nine runtime exports against a next-largest of three -- while eight of its
nine consumers imported it purely for behaviour.

Split along that line. utils/edition.ts now holds the slug constants,
EditionSlug, EditionCapabilities, the registry and the three predicates.
types/edition.ts keeps only the Edition interface, which is the shape of a
public.editions row and belongs with the other row types.

Pure move: the logic is byte-identical to the previous file's first 106 lines,
and the only other change is the import specifier at the eight behaviour call
sites. components/edition-select.tsx still imports Edition from types/edition
and is untouched.
